requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.6
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.3.1
python-musixmatch==1.0.0
requests==2.18.4
six==1.11.0
typed-ast==1.4.0
urllib3==1.22
Werkzeug==0.15.5
wrapt==1.11.2
